Minor Progress on Energy Shield
It blocks now, but throws a ton of erros. After I fix the errors I'll have to set up all the stuff with the cooldown, skill stuff, the take damage hook, making the material transparency dynamic and all that. Only pushing because I got to step away for a while.