We're on the road bois
didn't double-check all the changes like I usually do, but as it stands it's not falling apart so we should be good. Finalized new rig in blend (which entails a shit load more than this one sentence describes). ready to be animated again hopefully. and boy does it need to be animated again